Add files via upload
This is a revised library of my PureData patch which contains *only* the files needed to turn X/Y data into a wave file or to display it on an oscilloscope. Also included is a folder of X and Y coordinate files that I have made and can verify they work.